Move forwarding code to a package for reuse
